+++ src/client/components/TenantsList/TenantsList.js
@@ -86,7 +86,7 @@ Template.TenantsList.events({
             ...this,
             mdBody: 'TenantEditPanel',
             mdButtons: [ Modal.C.Button.CANCEL, Modal.C.Button.OK ],
-            mdClasses: 'modal-xl',
+            mdClasses: this.mdClasses || 'modal-xl',
             mdClassesContent: TenantsManager.configure().classes,
             mdTitle: pwixI18n.label( I18N, 'edit.modal_title' ),
             item: TenantsManager.list.byEntity( data.item._id )