+++ src/common/collections/records/server/functions.js
@@ -20,9 +20,11 @@ Records.server = {};
  */
 Records.server.getBy = async function( selector, userId ){
     check( selector, Object );
-    check( userId, String );
-    if( !await TenantsManager.isAllowed( 'pwix.tenants_manager.records.fn.get_by', userId, selector )){
-        return null;
+    if( userId ){
+        check( userId, String );
+        if( !await TenantsManager.isAllowed( 'pwix.tenants_manager.records.fn.get_by', userId, selector )){
+            return null;
+        }
     }
     const res = await Records.collection.find( selector ).fetchAsync();
     //console.debug( 'records', selector, res );