+++ ChangeLog.md
@@ -1,4 +1,4 @@
-# pwix:accounts-limanagerst
+# pwix:tenants-managers
 
 ## ChangeLog
 
@@ -8,6 +8,7 @@
 
     - Introduce 'pwix_tenants_manager_selecting' publication to let a user select a tenant and a validity period
     - Let both entity and record tabbed views be extended, thus bumping minor candidate version number
+    - Let the edition modal size be provided by the caller, defaulting to 'modal-xl'
 
 ### 1.0.0