+++ ChangeLog.md
@@ -9,6 +9,7 @@
     - Introduce 'pwix_tenants_manager_selecting' publication to let a user select a tenant and a validity period
     - Let both entity and record tabbed views be extended, thus bumping minor candidate version number
     - Let the edition modal size be provided by the caller, defaulting to 'modal-xl'
+    - Let the caller extends the buttons displayed on the right side of each row
 
 ### 1.0.0